+++ api/MailSender.py
@@ -7,22 +7,26 @@ from django.template.loader import render_to_string
 
 
 class EmailThread(threading.Thread):
-    def __init__(self, subject, body, from_email, recipient_list, fail_silently, html, bcc, reply_to):
+    def __init__(self, subject, body, from_email, recipient_list, fail_silently, html, reply_to,
+                 categories=None):
         self.subject = subject
         self.body = body
         self.recipient_list = recipient_list
         self.from_email = from_email
         self.fail_silently = fail_silently
         self.html = html
-        self.bcc = bcc
         self.reply_to = reply_to
+        self.categories = categories
         threading.Thread.__init__(self)
 
     def run(self):
-        msg = EmailMultiAlternatives(self.subject, self.body, self.from_email, self.recipient_list, bcc=self.bcc,
+        msg = EmailMultiAlternatives(self.subject, self.body, self.from_email, self.recipient_list,
                                      reply_to=[self.reply_to])
         if self.html:
             msg.attach_alternative(self.html, "text/html")
+        if self.categories:
+            msg.categories = self.categories
+
         msg.send(self.fail_silently)
 
 
@@ -47,18 +51,27 @@ class MailSender:
         subject = render_to_string(subject_file).strip()
         from_email = settings.DEFAULT_EMAIL_FROM
         to = target_email
-        bcc_email = settings.DEFAULT_EMAIL_BCC
         text_content = render_to_string(txt_file, template_ctxt)
         html_content = render_to_string(html_file, template_ctxt)
         reply_to = config('DEFAULT_REPLY_TO', default=None)
+        sanitized_team_name = template_ctxt.get('team').complete_name() \
+            .replace("ä", "ae").replace("ö", "oe").replace("ü", "ue").replace("ß", "ss") \
+            if 'team' in template_ctxt.keys() else None
+        categories = [
+            self.prefix,
+            template_prefix,
+            sanitized_team_name
+        ]
 
         if settings.SEND_EMAIL_ASYNC:
             EmailThread(subject, text_content, from_email, recipient_list=to, fail_silently=False, html=html_content,
-                        bcc=[bcc_email], reply_to=reply_to).start()
+                        reply_to=reply_to, categories=categories).start()
         else:
-            msg = EmailMultiAlternatives(subject, text_content, from_email, to, bcc=[bcc_email], reply_to=[reply_to])
+            msg = EmailMultiAlternatives(subject, text_content, from_email, to, reply_to=[reply_to])
             if html_content:
                 msg.attach_alternative(html_content, "text/html")
+            if categories:
+                msg.categories = categories
             msg.send()
 
     def send_email(self, prefix, email, data):